Play sound explicitly instead of in notification
For some reason, at some point toast notifications stop appearing as a popup in Windows. This makes them go unnoticed. By at least playing the sound by itself instead of in the notification, the user can hear something

Note that even Chrome and Mattermost lose popups. It is not a problem of this application exclusively